+++ app/Http/Controllers/ProjectTaskController.php
@@ -2,10 +2,12 @@
 
 namespace CodeProject\Http\Controllers;
 
+use CodeProject\Services\ProjectService;
 use CodeProject\Services\ProjectTaskService;
 use Illuminate\Http\Request;
 
 use CodeProject\Http\Requests;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectTaskController extends Controller
 {
@@ -13,13 +15,19 @@ class ProjectTaskController extends Controller
      * @var ProjectTaskService
      */
     private $service;
+    /**
+     * @var ProjectService
+     */
+    private $projectService;
 
     /**
      * @param ProjectTaskService $service
+     * @param ProjectService $projectService
      */
-    public function __construct(ProjectTaskService $service)
+    public function __construct(ProjectTaskService $service, ProjectService $projectService)
     {
         $this->service = $service;
+        $this->projectService = $projectService;
     }
 
     /**
@@ -30,6 +38,10 @@ class ProjectTaskController extends Controller
      */
     public function index($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->all($id);
     }
 
@@ -37,11 +49,19 @@ class ProjectTaskController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  Request $request
+     * @param int $projectId
      * @return Response
      */
-    public function store(Request $request)
+    public function store(Request $request, $projectId)
     {
-        return $this->service->create($request->all());
+        if ($this->checkProjectPermissions($projectId) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
+        $data = $request->all();
+        $data['project_id'] = $projectId;
+
+        return $this->service->create($data);
     }
 
     /**
@@ -53,6 +73,10 @@ class ProjectTaskController extends Controller
      */
     public function show($id, $taskId)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->find($id, $taskId);
     }
 
@@ -66,6 +90,10 @@ class ProjectTaskController extends Controller
      */
     public function update(Request $request, $id, $taskId)
     {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->update($request->all(), $taskId);
     }
 
@@ -78,6 +106,29 @@ class ProjectTaskController extends Controller
      */
     public function destroy($id, $taskId)
     {
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->service->delete($taskId);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        return $this->projectService->isOwner($projectId, Authorizer::getResourceOwnerId());
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        return $this->projectService->hasMember($projectId, Authorizer::getResourceOwnerId());
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
     }
 }
